requirements.txt
asgiref==3.6.0
black==22.12.0
click==8.1.3
Django==4.1.5
djangorestframework==3.14.0
isort==5.11.4
mypy-extensions==0.4.3
pathspec==0.11.0
phonenumbers==8.13.5
platformdirs==2.6.2
pytz==2022.7.1
sqlparse==0.4.3
tomli==2.0.1
